backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.9.0
pydantic-settings>=2.4.0
python-dotenv>=1.0.1
requests>=2.31.0
confluent-kafka>=2.5.0
google-generativeai>=0.7.0
httpx>=0.27.0
python-multipart>=0.0.9
